pyproject.toml
[project]
name = "order_service"
version = "0.1.0"
description = "Test task from FunTech"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.125.0",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
